+++ Krimpatul/src/BaseAbilities.cpp
@@ -1,10 +1,9 @@
-#include "Enumerators.hpp"
 #include "krpch.hpp"
 
 #include "BaseAbilities.hpp"
 
 // We set everything to zero, since they are going to change depending on class and level.
-BaseAbilities::BaseAbilities(ClassEnum cclass, unsigned short level)
+BaseAbilities::BaseAbilities(ClassEnum cclass, int level)
     : m_base_attack{0}, m_base_save{{{"FORT", 0}, {"REF", 0}, {"WILL", 0}}}
 {
     setBaseAttackBonus(cclass, level);
@@ -16,7 +15,7 @@ BaseAbilities::BaseAbilities(ClassEnum cclass, unsigned short level)
 // High: level of class.
 // Medium: level of class / ( 3 / 4 ) rounded down.
 // Low: level of class / 2 rounded down.
-auto BaseAbilities::setBaseAttackBonus(const ClassEnum cclass, const unsigned short level) -> void
+auto BaseAbilities::setBaseAttackBonus(const ClassEnum cclass, const int level) -> void
 {
     switch(cclass)
     {
@@ -40,7 +39,7 @@ auto BaseAbilities::setBaseAttackBonus(const ClassEnum cclass, const unsigned sh
 
 // Base saves are calculated according to class and level. For the formula, refer to the appropriate
 // function. Whether a class has high or low saves is explained in the Players Handbook.
-auto BaseAbilities::setBaseSaveBonus(const ClassEnum cclass, const unsigned short level) -> void
+auto BaseAbilities::setBaseSaveBonus(const ClassEnum cclass, const int level) -> void
 {
     switch(cclass)
     {
@@ -86,7 +85,7 @@ auto BaseAbilities::setBaseSaveBonus(const ClassEnum cclass, const unsigned shor
     }
 }
 
-auto BaseAbilities::getBaseAttackBonus() const -> unsigned short
+auto BaseAbilities::getBaseAttackBonus() const -> int
 {
     return m_base_attack;
 }
@@ -96,22 +95,20 @@ auto BaseAbilities::getBaseSaveBonus() const -> ArrayOfPairs
     return m_base_save;
 }
 
-auto BaseAbilities::getOneSave(unsigned short position) const
-    -> std::pair<std::string, unsigned short>
+auto BaseAbilities::getOneSave(int position) const -> std::pair<std::string, int>
 {
     return m_base_save.at(position);
 }
 
 // The formula for high base saves is level / 2 + 2 rounded down.
-auto BaseAbilities::calculateBaseSaveHigh(const unsigned short level) -> unsigned short
+auto BaseAbilities::calculateBaseSaveHigh(const int level) -> int
 {
-    return static_cast<unsigned short>(
-        std::floor(static_cast<double>(level) / consts::saves::form2) + consts::saves::form2);
+    return static_cast<int>(std::floor(static_cast<double>(level) / consts::saves::form2) +
+                            consts::saves::form2);
 }
 
 // The formula for low base saves is level / 2 rounded down.
-auto BaseAbilities::calculateBaseSaveLow(const unsigned short level) -> unsigned short
+auto BaseAbilities::calculateBaseSaveLow(const int level) -> int
 {
-    return static_cast<unsigned short>(
-        std::floor(calculateBaseSaveHigh(level) / consts::saves::form2));
+    return static_cast<int>(std::floor(calculateBaseSaveHigh(level) / consts::saves::form2));
 }
