+++ Krimpatul/src/BaseCharacter.hpp
@@ -1,7 +1,6 @@
 #ifndef BASECHARACTER_HPP
 #define BASECHARACTER_HPP
 
-#include "Constants.hpp"
 #include "krpch.hpp"
 
 #include "Abilities.hpp"
@@ -10,10 +9,6 @@
 #include "Race.hpp"
 #include "CClass.hpp"
 #include "Hitpoints.hpp"
-#include <array>
-#include <c++/9.2.1/x86_64-pc-linux-gnu/bits/c++config.h>
-#include <string>
-#include <utility>
 
 class BaseCharacter : public Race, public CClass, public Alignment
 {
@@ -27,23 +22,23 @@ public:
     auto operator=(BaseCharacter &&) -> BaseCharacter & = delete;
     /************************/
     /* Setters */
-    auto setAttackBonusMelee(BaseAbilities /*bab*/, Abilities /*str*/, Race /*size*/,
-                             short /*misc*/) -> void;
+    auto setAttackBonusMelee(BaseAbilities /*bab*/, Abilities /*str*/, Race /*size*/, int /*misc*/)
+        -> void;
     auto setAttackBonusRanged(BaseAbilities /*bab*/, Abilities /*dex*/, Race /*size*/,
-                              short /*penalty*/, short /*misc*/) -> void;
+                              int /*penalty*/, int /*misc*/) -> void;
     auto setSaves(std::array<int, consts::saves::count> /*values*/) -> void;
-    auto setOneSave(unsigned short /*position*/, int /*value*/) -> void;
-    auto setSpellResistance(unsigned short /*sr*/) -> void;
-    auto setInitiative(short /*dex_mod*/, short /*misc*/) -> void;
-    auto setGrapple(short /*bab*/, short /*str_mod*/, short /*size_mod*/, short /*misc*/) -> void;
+    auto setOneSave(int /*position*/, int /*value*/) -> void;
+    auto setSpellResistance(int /*sr*/) -> void;
+    auto setInitiative(int /*dex_mod*/, int /*misc*/) -> void;
+    auto setGrapple(int /*bab*/, int /*str_mod*/, int /*size_mod*/, int /*misc*/) -> void;
     /************************/
     /* Getters */
     [[nodiscard]] auto getAttackBonusMelee() const -> int;
     [[nodiscard]] auto getAttackBonusRanged() const -> int;
     [[nodiscard]] auto getSaves() const -> ArrayOfPairs;
-    [[nodiscard]] auto getOneSave(unsigned short /*position*/) const -> int;
-    [[nodiscard]] auto getSpellResistance() const -> unsigned short;
-    [[nodiscard]] auto getInitiative() const -> short;
+    [[nodiscard]] auto getOneSave(int /*position*/) const -> int;
+    [[nodiscard]] auto getSpellResistance() const -> int;
+    [[nodiscard]] auto getInitiative() const -> int;
     [[nodiscard]] auto getGrapple() const -> int;
     /************************/
 protected:
@@ -55,11 +50,11 @@ private:
     BaseAbilities  m_base_abilities;
     HitPoints      m_hitpoints;
     ArrayOfPairs   m_saves;
-    unsigned short m_level{1};
+    int            m_level{1};
     int            m_attack_bonus_melee{0};
     int            m_attack_bonus_ranged{0};
-    unsigned short m_spell_resistance{0};
-    short          m_initiative{0};
+    int            m_spell_resistance{0};
+    int            m_initiative{0};
     int            m_grapple{0};
 };
